+++ lib/src/tabs/tabs.dart
@@ -3,3 +3,4 @@ library tabs;
 import 'package:flutter/material.dart';
 
 part 'src/tab.dart';
+part 'src/view.dart';